requirements.txt
python-dotenv
pydantic
matplotlib
numpy
pandas
torch
wandb
lightning
scikit-learn
